% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generating_functions.R
\name{BM}
\alias{BM}
\title{Brownian Motion}
\usage{
BM(n, t, sigma = 1)
}
\arguments{
\item{n}{TODO}

\item{t}{TODO}

\item{sigma}{TODO}
}
\value{
TODO
}
\description{
Brownian Motion
}
\examples{
0 # TODO
}
