% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generating_functions.R
\name{cov.ou}
\alias{cov.ou}
\title{Covariance kernel of Ornstein-Uhlenbeck process}
\usage{
cov.ou(t, sigma, theta, sigma0 = 0)
}
\arguments{
\item{t}{2 dimensional}

\item{sigma}{influence of brownian motion to process}

\item{theta}{mean reversion}

\item{sigma0}{Variance of starting distribution, 0 if X0 is deterministic}
}
\value{
value of covariance
}
\description{
Covariance kernel of Ornstein-Uhlenbeck process
}
\examples{
0 # TODO
}
