% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.r
\name{epak_2d}
\alias{epak_2d}
\title{Two-Dimensional Epanechnikov Kernel}
\usage{
epak_2d(x, h = 1)
}
\arguments{
\item{x}{Two dimensional vector at which the kernel shall be evaluated.}

\item{h}{Bandwidth parameter for nonparametric regression. Default is h = 1.}
}
\value{
Real value between 0 and 1.
}
\description{
Two dimensional Epanechnikov Kernel with optional bandwidth parameter.
}
\examples{
epak_2d(c(0.2, 0.13), 0.4)
}
