% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_polynomial_weights.R
\name{eval_weights}
\alias{eval_weights}
\title{Evaluates the weights from local_polynomial_weights with data}
\usage{
eval_weights(W, Z)
}
\arguments{
\item{W}{weights object from local_polynomial_weights}

\item{Z}{transformed data}
}
\value{
Matrix/Array with evaluation
}
\description{
Evaluates the weights from local_polynomial_weights with data
}
\examples{
0 # TODO
}
