% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generating_functions.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generates values to test functions}
\usage{
generate_test_data(...)
}
\arguments{
\item{...}{grid type}
}
\value{
values
}
\description{
Generates values to test functions
}
\examples{
generate_test_data()
}
