% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.r
\name{invert3x3}
\alias{invert3x3}
\title{Inverts a 3x3 matrix}
\usage{
invert3x3(M, eps = 10^(-6))
}
\arguments{
\item{M}{Matrix of Dimension 3x3}

\item{eps}{If the the determinant of M is smaller than this value the function returns an error (unstable matrix inversion).}
}
\value{
The inverse 3x3 matrix.
}
\description{
Inverts a 3x3 matrix more quickly than solve.
}
\examples{
M = matrix(c(2,3,5,2,5,3,5,3,2), 3, 3)
M_inv = invert3x3(M)
round(M \%*\% M_inv, 15)
}
