% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{k_fold_cv}
\alias{k_fold_cv}
\title{k_fold_c}
\usage{
k_fold_cv(
  Y,
  h.seq,
  K = 2,
  m = 1,
  h.parallel = F,
  h.parallel.environment = F,
  ...
)
}
\arguments{
\item{Y}{Raw observations without any transformation. Numerical with dimension n x p.}

\item{h.seq}{Sequence of bandwidth to compare in cross validation. Numerical vector.}

\item{K}{Into how many subset shall the data be devided?}

\item{m}{Degree of local polynomial estimation. 1 and 2 is possible.}

\item{h.parallel}{Logical value wether the}

\item{h.parallel.environment}{Logical value that indicates if the 'makeCluster', 'plan' and 'stopCluster' functions shall be called}

\item{...}{Further arguments passend to 'local_polynomial_weights'.}
}
\value{
Bandwidth with minimal sup error in cross validation
}
\description{
K-Fold Cross Validation
}
\examples{
0 # TODO:
}
