% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_polynomial_weights.R
\name{local_polynomial_weights}
\alias{local_polynomial_weights}
\title{Calculated the weights of the linear local polynomial estimator}
\usage{
local_polynomial_weights(
  p,
  h,
  p.eval,
  parallel = F,
  m = 1,
  del = 0,
  x.design.grid = NULL,
  grid.type = "less",
  eval.type = "full",
  ...
)
}
\arguments{
\item{p}{amount of design (observation) points on one axis}

\item{h}{bandwidth parameter}

\item{p.eval}{amount of evaluation points (can be NULL if `eval.type = diagonal`)}

\item{parallel}{Paralization via the future package}

\item{m}{`m = 0`refers to Nadaraya-Watson, `m = 1` to the local linear estimator and `m = 2` to local quadratic}

\item{del}{Calculation of derivatives}

\item{x.design.grid}{Can be used for custom designs}

\item{grid.type}{which kind of grid (`less`, `without diagonal`, `full`)}

\item{eval.type}{at all evaluation points or only at the diagonal (`full`, `diagonal`)}

\item{...}{further arguments}
}
\value{
TODO: write something usefull -> can we create an object here?
}
\description{
according to paper Berger/Holzmann (2024)
}
\examples{
local_polynomial_weights(20, 0.3, 25, m = 2, del = 0)
}
