% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{lopocv}
\alias{lopocv}
\title{lopocv}
\usage{
lopocv(Y, h.seq, m = 2, h.parallel = F, h.parallel.environment = F, ...)
}
\arguments{
\item{Y}{Raw observations without any transformation. Numerical with dimension n x p.}

\item{h.seq}{Sequence of bandwidth to compare in cross validation. Numerical vector.}

\item{m}{Degree of local polynomial estimation. 1 and 2 is possible.}

\item{h.parallel}{Logical value}

\item{h.parallel.environment}{Logical value that indicates if the 'makeCluster', 'plan' and 'stopCluster' functions shall be called}

\item{...}{Further arguments passend to 'local_polynomial_weights' such as 'parallel' and 'parallel.environment'}
}
\value{
Bandwidth with minimal sup error in cross validation
}
\description{
Leave One Plane Out Cross Validation.
}
\examples{
0 # TODO:
}
