% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generating_functions.R
\name{mu}
\alias{mu}
\title{Mean Function}
\usage{
mu(x)
}
\arguments{
\item{x}{vector with points to evaluate function}
}
\value{
vector of same length as x
}
\description{
Mean Function
}
\examples{
mu(0.3)
}
