% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.r
\name{observation_grid}
\alias{observation_grid}
\title{Creates two dimensional observation grids}
\usage{
observation_grid(p = NULL, x = NULL, comp = "less")
}
\arguments{
\item{p}{Amount of design points.}

\item{x}{Needed if p is not given. p-dimensional vector with design points on a real interval. If x
is not given x is created with p equidistant design points.}

\item{comp}{Indicates the structure of the grid:

-- `less`: The first coordinate from the grid is always strictly smaller then the second, `gtr`: vice versa.

-- `lesseq`:  The first coordinate from the grid is always smaller or equal then the second, `gtreq`: vice versa.

-- `without diagonal`: Elements with equal values are excluded from the grid.

-- `full`: No element is excluded.}
}
\value{
Matrix of dimension q x 2 where q depends on comp. Note the rownames of the matrix refer to the initial
numbering before excluding certain values as indicated by comp.
}
\description{
Creates two dimensional observation grids.
}
\examples{
observation_grid(10)
}
