% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.r
\name{observation_transformation}
\alias{observation_transformation}
\title{Transforms observations according to Berger/Holzmann (2024)}
\usage{
observation_transformation(Y, grid.type = "less", na.rm = F)
}
\arguments{
\item{Y}{Observations in Form of a n x p matrix}

\item{grid.type}{which kind of observation transformation (`less`, `without diagonal`, `full`)}
}
\value{
Vector with transformed obseravtions of dimension `p(p-1)/2` or `p(p-1)` or `p^2`
}
\description{
Transforms observations according to Berger/Holzmann (2024)
}
\examples{
0
}
