% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generating_functions.R
\name{OU}
\alias{OU}
\title{Ornstein Uhlenbeck Process}
\usage{
OU(n, t = seq(0, 1, len = 201), mu = 0, alpha = 1, sigma = 1, x0 = 0)
}
\arguments{
\item{n}{how many processes}

\item{t}{points of evaluation}

\item{mu}{mean}

\item{alpha}{TODO}

\item{sigma}{TODO}

\item{x0}{TODO}
}
\value{
TODO
}
\description{
Uses the ... package !!
}
\examples{
0 # TODO
}
