% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weigths.r
\name{slice_matrix}
\alias{slice_matrix}
\title{Slices a Matrix into an Array}
\usage{
slice_matrix(M, d1, d2)
}
\arguments{
\item{M}{Matrix that shall be sliced}

\item{d1}{rows of matrices}

\item{d2}{how many matrices shall be created}
}
\value{
An Array of dimension d1 x d2 x d3, where d3 is the amount of rows of M
}
\description{
Slices a Matrix into an Array according to the dimension of the matrix. Needed to handle the return of applying the
 weights.point function
}
\examples{
slice_matrix(M = matrix(1:27, 9, 3), d1 = 3, d2 = 3)
}
