% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.r
\name{U}
\alias{U}
\title{Vector U for bivariate local polynomial estimator}
\usage{
U(x, h, m = 1)
}
\arguments{
\item{x}{two dimensional vector of evaluation.}

\item{h}{bandwidth parameter.}

\item{m}{`m = 0`refers to Nadaraya-Watson, `m = 1` to the local linear estimator and `m = 2` to local quadratic}
}
\value{
one, three or six dimensional vector (depending on `m`)
}
\description{
Vector U for bivariate local polynomial estimator
}
\examples{
U(rep(0.5, 6), h = 0.25, m = 2)
}
