% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generating_functions.R
\name{z_2rv}
\alias{z_2rv}
\title{z_2rv}
\usage{
z_2rv(n = 1, p = 100, t = NULL, a = 2/3, b = sqrt(2) * 2/3, c = 1.25)
}
\arguments{
\item{n}{Amount of observated curves in FDA setting}

\item{p}{Amount of Design points in FDA setting.}

\item{t}{Vector for design points if the design shall differ from the equidistant design. If not provided equidistant setting is used.}

\item{a}{Parameter}

\item{b}{Parameter}

\item{c}{Parameter}
}
\value{
Matrix of dimension n x p.
}
\description{
z_2rv
}
\examples{
z_2rv(n = 10)
}
