% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_polynomial_weights.R
\name{eval_weights}
\alias{eval_weights}
\title{Evaluates the weights from local_polynomial_weights with data}
\usage{
eval_weights(W, Z)
}
\arguments{
\item{W}{weights object from local_polynomial_weights}

\item{Z}{transformed data}
}
\value{
Matrix/Array with evaluation
}
\description{
Evaluates weights for correctly transformed weights. If evaluation points of W
are given via "without.diagonal" also the observations need to be transformed with "grid.type = without.diagonal".
}
\examples{
0 # TODO
}
